requirements.txt
fastapi
Pillow
numpy
tensorflow-cpu
Keras
uvicorn
prometheus-client
prometheus-fastapi-instrumentator
psutil
